Send scroll_id inside a json rather that text
It's likely to cause issues with elastic 5.3 as Content-Type handling becomes
more and more strict.

eg.:
curl -H"Content-Type: application/json"  -XGET localhost:9200/_search/scroll?scroll=1m -d 'DXF...'
now fails with 5.3.x Failed to parse request body.
Elastic 5.1 was lenient and accepted this body.
Since we default Content-Type to application/json this is likely cause issues
in Elastica as well.
Using an array instead of plain string should force the use of a JSON body which coherent
with the content-type we set.